Finished refactoring frontend file structure
