+++ data.json
@@ -47,6 +47,18 @@
                         "starting_xp": "0",
                         "goal_xp": "3"
                     }
+                },
+                "3022146279385490229656229347854996392844497171979249865727464850011966461481": {
+                    "type": "XP",
+                    "title": "Gain 33 XP by 13/04/2024",
+                    "end_date": "13/04/2024",
+                    "status": "in progress",
+                    "fail_date": "",
+                    "data": {
+                        "start_date": "12/04/2024",
+                        "starting_xp": "0",
+                        "goal_xp": "33"
+                    }
                 }
             },
             "flashcards": {
