+++ frontend/src/componments/GridContainer/index.js
@@ -1,0 +1,1 @@
+export {default} from './GridContainer';
