+++ frontend/src/componments/GridItem/GridItem.css
