+++ frontend/src/componments/GridItem/GridItem.js
