+++ frontend/src/componments/GridItem/index.js
@@ -1,0 +1,1 @@
+export {default} from './GridItem';
