+++ frontend/src/containers/BentoPanel/index.js
@@ -1,0 +1,1 @@
+export {default} from './BentoPanel';
