+++ frontend/src/containers/BlobBackground/BlobBackground.css
