+++ frontend/src/containers/FlashcardOverview/FlashcardFolder/FlashcardFolder.js
@@ -3,8 +3,8 @@ import {motion} from 'framer-motion';
 import Image from '../../../componments/Image/Image';
 import Paragraph from '../../../componments/Text/Paragraph/Paragraph';
 import Heading5 from '../../../componments/Text/Heading5/Heading5';
-import GridContainer from '../../GridContainer';
-import GridItem from '../../GridItem';
+import GridContainer from '../../../componments/GridContainer/GridContainer';
+import GridItem from '../../../componments/GridItem/GridItem';
 import ReviewBarChart from '../../ReviewBarChart';
 import Button from '../../../componments/Button';
 import { easeIn } from '../../../animations/animations';
