+++ frontend/src/containers/FlashcardOverview/FlashcardItem/FlashcardItem.js
@@ -3,10 +3,10 @@ import Heading5 from '../../../componments/Text/Heading5/Heading5';
 import Paragraph from '../../../componments/Text/Paragraph/Paragraph';
 import Button from '../../../componments/Button';
 import Image from '../../../componments/Image/Image';
-import GridContainer from '../../GridContainer';
-import GridItem from '../../GridItem';
+import GridContainer from '../../../componments/GridContainer/GridContainer';
+import GridItem from '../../../componments/GridItem/GridItem';
 import ReviewBarChart from '../../ReviewBarChart';
-import CardOperationsPopup from '../../CardOperationsPopup';
+import CardOperationsPopup from '../../Modal/CardOperationsPopup/CardOperationsPopup';
 
 import circledTick from '../../../static/circled-tick.svg';
 import emptyCircle from '../../../static/empty-circle.svg';
