+++ frontend/src/containers/MobileSidePanel/MobileSidePanel.css
