+++ frontend/src/containers/MobileSidePanel/index.js
@@ -1,0 +1,1 @@
+export {default} from './MobileSidePanel';
