+++ frontend/src/containers/Modal/CardOperationsPopup/CardOperationsPopup.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import MenuItem from './MenuItem';
+import MenuItem from '../../MenuItem';
 
-import renameIcon from '../static/rename-icon.svg';
-import moveIcon from '../static/move-icon.svg';
-import deleteIcon from '../static/bin-icon.svg';
+import renameIcon from '../../../static/rename-icon.svg';
+import moveIcon from '../../../static/move-icon.svg';
+import deleteIcon from '../../../static/bin-icon.svg';
 
 import './CardOperationsPopup.css'
