+++ frontend/src/containers/Modal/CardOperationsPopup/index.js
@@ -1,0 +1,1 @@
+export {default} from './CardOperationsPopup';
