+++ frontend/src/containers/Modal/EditGoalPopup/index.js
@@ -1,0 +1,1 @@
+export {default} from './EditGoalPopup';
