+++ frontend/src/containers/Modal/MoveFolderDialogue/index.js
@@ -1,0 +1,1 @@
+export {default} from './MoveFolderDialogue';
