+++ frontend/src/containers/Modal/NewGoalPopup/index.js
@@ -1,0 +1,1 @@
+export {default} from './NewGoalPopup';
