+++ frontend/src/containers/Modal/index.js
@@ -1,0 +1,3 @@
+export { default as CardOperationsPopup } from './CardOperationsPopup';
+export { default as EditGoalPopup } from './EditGoalPopup';
+export {default as MoveFolderDialogue } from './MoveFolderDialogue';
