+++ frontend/src/containers/SidePanel/SidePanel.css
