+++ frontend/src/containers/SidePanel/index.js
@@ -1,0 +1,1 @@
+export {default} from './SidePanel';
