+++ frontend/src/containers/StreakWidget/StreakWidget.js
@@ -1,12 +1,11 @@
 import {React, useState, useEffect} from 'react';
-import DelayedElement from './DelayedElement';
-import Heading5 from '../componments/Text/Heading5/Heading5';
-import WhiteOverlay from '../componments/WhiteOverlay/WhiteOverlay';
-import Header from '../componments/Text/Header/Header';
-import apiManager from '../api/Api';
-import Heading4 from '../componments/Text/Heading4/Heading4';
-import { getCookie } from '../api/Authentication';
-import '../App.css';
+import DelayedElement from '../DelayedElement';
+import Heading5 from '../../componments/Text/Heading5/Heading5';
+import WhiteOverlay from '../../componments/WhiteOverlay/WhiteOverlay';
+import Header from '../../componments/Text/Header/Header';
+import apiManager from '../../api/Api';
+import Heading4 from '../../componments/Text/Heading4/Heading4';
+import { getCookie } from '../../api/Authentication';
 
 function StreakWidget () {
     const [streak, setStreak] = useState(null)
