+++ frontend/src/containers/StreakWidget/index.js
@@ -1,0 +1,1 @@
+export {default} from './StreakWidget';
