+++ frontend/src/screens/Flashcards.js
@@ -2,9 +2,9 @@ import {React, useState, useEffect} from 'react';
 import { Helmet } from 'react-helmet';
 import '../App.css';
 import BlobBackground from '../containers/BlobBackground';
-import GridContainer from '../containers/GridContainer';
-import GridItem from '../containers/GridItem';
-import SidePanel from '../containers/SidePanel';
+import GridContainer from '../componments/GridContainer/GridContainer';
+import GridItem from '../componments/GridItem/GridItem';
+import SidePanel from '../containers/SidePanel/SidePanel';
 import WhiteOverlay from '../componments/WhiteOverlay/WhiteOverlay';
 import HamburgerBar from '../containers/HamburgerBar/HamburgerBar';
 import FlashcardOverview from '../containers/FlashcardOverview/FlashcardOverview';
@@ -13,7 +13,7 @@ import Button from '../componments/Button';
 import GhostButton from '../componments/GhostButton';
 import SearchBar from '../componments/SearchBar/SearchBar';
 import ReviewBarChartKey from '../containers/ReviewBarChartKey/ReviewBarChartKey';
-import MoveFolderDialogue from '../containers/MoveFolderDialogue';
+import MoveFolderDialogue from '../containers/Modal/MoveFolderDialogue/MoveFolderDialogue';
 import apiManager from '../api/Api';
 import '../componments/Text/Text/Text.css';
 import '../componments/Text/Link/Link.css';
