+++ frontend/src/screens/SignIn.js
@@ -4,15 +4,15 @@ import '../App.css';
 import firebase from 'firebase/compat/app';
 import { signInWithGoogle } from '../api/Authentication';
 import BlobBackground from '../containers/BlobBackground';
-import GridContainer from '../containers/GridContainer';
-import GridItem from '../containers/GridItem';
+import GridContainer from '../componments/GridContainer/GridContainer';
+import GridItem from '../componments/GridItem/GridItem';
 import Header from '../componments/Text/Header/Header';
 import Subheader from '../componments/Text/Subheader/Subheader';
 import Image from '../componments/Image/Image';
 import Paragraph from '../componments/Text/Paragraph/Paragraph';
 import Button from '../componments/Button';
 import DeveloperSection from '../componments/DeveloperSection';
-import BentoPanel from '../containers/BentoPanel';
+import BentoPanel from '../containers/BentoPanel/BentoPanel';
 import MailChimpWidget from '../containers/MailChimpWidget';
 import goals from '../static/goals.png';
 import heatmap from '../static/heatmap.png';
